+++ example/modular_di/lib/main.dart
@@ -1,0 +1,11 @@
+import 'package:force_inject/force_inject.dart';
+import 'modules/auth_module.dart';
+import 'services/user_service.dart';
+
+void main() {
+  final services = ServiceCollection()..addModule(AuthModule());
+  final provider = services.buildServiceProvider();
+
+  final userService = provider.get<UserService>();
+  print(userService.userData);
+}